continue to refactor
