+++ internal/redact/types.go
@@ -1,0 +1,7 @@
+package redact
+
+type RedactedMeta struct {
+	Location string
+	Size     int
+	Type     string
+}
